refactor: move config package to api
