+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/api/config/AnnotationConfiguration.java
@@ -1,5 +1,5 @@
 
-package io.javaoperatorsdk.operator.config;
+package io.javaoperatorsdk.operator.api.config;
 
 import java.util.Set;
