+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/api/config/ControllerConfiguration.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.config;
+package io.javaoperatorsdk.operator.api.config;
 
 import java.util.Collections;
 import java.util.Set;
