+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/retry/GenericRetry.java
@@ -1,6 +1,6 @@
 package io.javaoperatorsdk.operator.processing.retry;
 
-import io.javaoperatorsdk.operator.config.RetryConfiguration;
+import io.javaoperatorsdk.operator.api.config.RetryConfiguration;
 
 public class GenericRetry implements Retry {
