+++ samples/spring-boot-plain/src/main/java/io/javaoperatorsdk/operator/sample/Config.java
@@ -1,5 +1,8 @@
 package io.javaoperatorsdk.operator.sample;
 
+import io.javaoperatorsdk.operator.Operator;
+import io.javaoperatorsdk.operator.api.ResourceController;
+import io.javaoperatorsdk.operator.api.config.DefaultConfigurationService;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.javaoperatorsdk.operator.Operator;
@@ -24,7 +27,7 @@ public class Config {
   //  Register all controller beans
   @Bean
   public Operator operator(KubernetesClient client, List<ResourceController> controllers) {
-    Operator operator = new Operator(client);
+    Operator operator = new Operator(client, DefaultConfigurationService.instance());
     controllers.forEach(c -> operator.registerControllerForAllNamespaces(c));
     return operator;
   }
