+++ spring-boot-starter/src/main/java/io/javaoperatorsdk/operator/springboot/starter/OperatorAutoConfiguration.java
@@ -13,11 +13,11 @@ import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.openshift.client.DefaultOpenShiftClient;
 import io.javaoperatorsdk.operator.Operator;
 import io.javaoperatorsdk.operator.api.ResourceController;
-import io.javaoperatorsdk.operator.config.AnnotationConfiguration;
-import io.javaoperatorsdk.operator.config.ClientConfiguration;
-import io.javaoperatorsdk.operator.config.ConfigurationService;
-import io.javaoperatorsdk.operator.config.ControllerConfiguration;
-import io.javaoperatorsdk.operator.config.RetryConfiguration;
+import io.javaoperatorsdk.operator.api.config.AnnotationConfiguration;
+import io.javaoperatorsdk.operator.api.config.ClientConfiguration;
+import io.javaoperatorsdk.operator.api.config.ConfigurationService;
+import io.javaoperatorsdk.operator.api.config.ControllerConfiguration;
+import io.javaoperatorsdk.operator.api.config.RetryConfiguration;
 import io.javaoperatorsdk.operator.processing.retry.GenericRetry;
 import io.javaoperatorsdk.operator.processing.retry.Retry;
 import org.apache.commons.lang3.StringUtils;
