+++ spring-boot-starter/src/main/java/io/javaoperatorsdk/operator/springboot/starter/OperatorProperties.java
@@ -2,7 +2,7 @@ package io.javaoperatorsdk.operator.springboot.starter;
 
 import java.util.Optional;
 
-import io.javaoperatorsdk.operator.config.ClientConfiguration;
+import io.javaoperatorsdk.operator.api.config.ClientConfiguration;
 
 public class OperatorProperties implements ClientConfiguration {
